moved files around for music playing
