+++ client/pages/index.tsx
@@ -6,19 +6,19 @@ import Footer from "./components/Footer";
 import PlayButton from "./components/PlayButton";
 import Link from "next/link";
 import { backend } from "./components/Constants";
-import song1 from "../../data/newSongs/ai-generated-7926621_640.jpg";
-import song2 from "../../data/newSongs/ai-generated-8716960_640.jpg";
-import song3 from "../../data/newSongs/universe-8418706_640.jpg";
-import song4 from "../../data/newSongs/walkTogether.jpg";
-import song5 from "../../data/newSongs/LoFi Chill.png";
-import product1 from "../../data/trendingProducts/duck.jpg";
-import product2 from "../../data/trendingProducts/sweater.jpg";
-import product3 from "../../data/trendingProducts/ring.jpg";
-import product4 from "../../data/trendingProducts/art.jpg";
-import product5 from "../../data/trendingProducts/teddyBear.jpg";
-import listen1 from "../../data/recentlyListened/riseAndShine.jpg";
-import listen2 from "../../data/recentlyListened/i will be here.png";
-import listen3 from "../../data/recentlyListened/OneStepFurther.png";
+import song1 from "../public/data/newSongs/ai-generated-7926621_640.jpg";
+import song2 from "../public/data/newSongs/ai-generated-8716960_640.jpg";
+import song3 from "../public/data/newSongs/universe-8418706_640.jpg";
+import song4 from "../public/data/newSongs/walkTogether.jpg";
+import song5 from "../public/data/newSongs/LoFi Chill.png";
+import product1 from "../public/data/trendingProducts/duck.jpg";
+import product2 from "../public/data/trendingProducts/sweater.jpg";
+import product3 from "../public/data/trendingProducts/ring.jpg";
+import product4 from "../public/data/trendingProducts/art.jpg";
+import product5 from "../public/data/trendingProducts/teddyBear.jpg";
+import listen1 from "../public/data/recentlyListened/riseAndShine.jpg";
+import listen2 from "../public/data/recentlyListened/i will be here.png";
+import listen3 from "../public/data/recentlyListened/OneStepFurther.png";
 
 interface Track {
   id: number;
